Enhance glottolog data (flexible options)

It is recommended to use glottobooster, but this function is more flexible in removing/adding columns.

glottoboosterflex(
  glottologdata = NULL,
  space = TRUE,
  addfamname = TRUE,
  addisolates = TRUE,
  rmdialects = TRUE,
  rmfamilies = TRUE,
  addfamsize = TRUE,
  addfamsizerank = TRUE,
  rmbookkeeping = TRUE,
  rmartifam = TRUE,
  rmsignlangs = TRUE,
  rename = TRUE
)

Arguments

glottologdata

data from glottolog, can be downloaded with glottoget("glottolog").

space

Return spatial object?

addfamname

Add column with familiy names?

addisolates

Add column to identify isolates?

rmdialects

Remove rows with dialects?

rmfamilies

Remove rows with families?

addfamsize

Add column with family size?

addfamsizerank

Add column with family size rank?

rmbookkeeping

Remove bookkeeping rows and delete bookkeeping column?

rmartifam

Remove rows that refer to artificial families?

rmsignlangs

Remove rows with sign languages?

rename

Rename columns "id" to "glottocode" and "iso639p3code" to "isocode"

Value

glottologdata object, either a spatial object (class: sf) or a data.frame.

Details

Note that the different options are additional. For example, if you set rmfamilies to TRUE and space = TRUE, most families will still be removed since they lack spatial coordinates. Another example, depending on whether artificial families are removed, the total number of families also increases/decreases.

See also

Examples

# \donttest{
glottologdata <- glottoget("glottolog")
glottologdata <- glottoboosterflex(glottologdata)
# }