Generate empty glottodata or glottosubdata for a set of glottocodes.

Creates glottodata/glottosubdata and optionally save it as excel file.

glottocreate(
  glottocodes,
  variables,
  meta = TRUE,
  filename = NULL,
  simplify = TRUE,
  groups = NULL,
  n = NULL,
  levels = NULL,
  check = FALSE,
  maintainer = NULL,
  email = NULL,
  citation = NULL,
  url = NULL
)

Arguments

glottocodes

Character vector of glottocodes

variables

Either a vector with variable names, or a single number indicating the total number of variable columns to be generated

meta

Should metatables be created?

filename

Optional name of excel file where to store glottodata

simplify

By default, if a glottodata table is created without metadata, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)

groups

Character vector of group names (only for glottosubdata)

n

Optional, number of records to be assigned to each group (only for glottosubdata)

levels

Optional character vector with levels across all variables

check

Should glottocodes be checked? Default is FALSE because takes much time to run.

maintainer

Name of the person/organization maintaining the data (optional, added to readme tab)

email

Email address of maintainer/contact person (optional, added to readme tab)

citation

How to cite the data (optional, added to readme tab)

url

Link to a webpage (optional, added to readme tab).

Value

A glottodata or glottosubdata object (either with or without metadata). The output can be a list or a data.frame.

Details

By default, glottodata will be created. In case a groups argument is provided, glottosubdata will be created.

glottodata has one table for all languages (and a number of metatables if meta = TRUE), with one row per glottocode. glottosubdata has one table for each language (and a number of metatables if meta = TRUE), with one row per glottosubcode.

Run glottoget("demodata") or glottoget("demosubdata") to see examples.

In case you already have your own dataset and want to convert it into glottodata, use: glottoconvert().

Examples

# Creates glottodata table without metadata tables
glottocreate(glottocodes = c("yucu1253", "tani1257"),
variables = 3, meta = FALSE)
#>   glottocode var001 var002 var003
#> 1   yucu1253     NA     NA     NA
#> 2   tani1257     NA     NA     NA

# Creates glottodata table with metadata tables (stored in a list):
glottocreate(glottocodes = c("yucu1253", "tani1257"), variables = 3)
#> $glottodata
#>   glottocode var001 var002 var003
#> 1   yucu1253     NA     NA     NA
#> 2   tani1257     NA     NA     NA
#> 
#> $structure
#>   varname type levels weight group subgroup
#> 1  var001   NA     NA      1    NA       NA
#> 2  var002   NA     NA      1    NA       NA
#> 3  var003   NA     NA      1    NA       NA
#> 
#> $description
#>   varname description reference remarks lev_Y lev_N lev_NA lev_A lev_B lev_C
#> 1  var001          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 2  var002          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 3  var003          NA        NA      NA    NA    NA     NA    NA    NA    NA
#>   lev_D
#> 1    NA
#> 2    NA
#> 3    NA
#> 
#> $references
#>   glottocode reference page var001_ref var001_page var002_ref var002_page
#> 1   yucu1253        NA   NA         NA          NA         NA          NA
#> 2   tani1257        NA   NA         NA          NA         NA          NA
#>   var003_ref var003_page
#> 1         NA          NA
#> 2         NA          NA
#> 
#> $remarks
#>   glottocode remark var001_remark var002_remark var003_remark
#> 1   yucu1253     NA            NA            NA            NA
#> 2   tani1257     NA            NA            NA            NA
#> 
#> $contributors
#>   glottocode contributor var001_contributor var002_contributor
#> 1   yucu1253          NA                 NA                 NA
#> 2   tani1257          NA                 NA                 NA
#>   var003_contributor
#> 1                 NA
#> 2                 NA
#> 
#> $sample
#>   glottocode group subgroup
#> 1   yucu1253    NA       NA
#> 2   tani1257    NA       NA
#> 
#> $readme
#>                                                        info           value
#> 1                                                maintainer            <NA>
#> 2                                                     email            <NA>
#> 3                                                  citation            <NA>
#> 4                                                       url            <NA>
#> 5 This database was created using the glottospace R package version 0.0.113
#> 
#> $lookup
#>   type_lookup
#> 1        symm
#> 2       asymm
#> 3      factor
#> 4     ordered
#> 5     numeric
#> 6    ordratio
#> 7    logratio
#>                                                                                                                                                   type_legend
#> 1 Symmetric binary: positive match is considered the same as a negative match. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 2          Asymmetric binary: positive match and negative match are not equal. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 3                                    Nominal variables. Levels should be given by a letter or abbreviation. For example: A, B, C. NA stands for missing value
#> 4                                                                                                              Ordinal variables. NA stands for missing value
#> 5                                                                                                Interval scaled variables. Can have any numeric value and NA
#> 6                                               Ratio scaled variables. Setting type to ordratio indicates that values should be treated as ordinal variables
#> 7                                     Ratio scaled variables. Setting type to logratio indicates that values should be log transformed in subsequent analyses
#> 
#> attr(,"class")
#> [1] "glottodata" "list"      

# Creates glottosubdata table (stored in a list)
glottocreate(glottocodes = c("yucu1253", "tani1257"),
variables = 3, groups = c("a", "b") )
#> $yucu1253
#>     glottosubcode var001 var002 var003
#> 1 yucu1253_a_0001     NA     NA     NA
#> 2 yucu1253_b_0001     NA     NA     NA
#> 
#> $tani1257
#>     glottosubcode var001 var002 var003
#> 1 tani1257_a_0001     NA     NA     NA
#> 2 tani1257_b_0001     NA     NA     NA
#> 
#> $structure
#>   varname type levels weight group subgroup
#> 1  var001   NA     NA      1    NA       NA
#> 2  var002   NA     NA      1    NA       NA
#> 3  var003   NA     NA      1    NA       NA
#> 
#> $description
#>   varname description reference remarks lev_Y lev_N lev_NA lev_A lev_B lev_C
#> 1  var001          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 2  var002          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 3  var003          NA        NA      NA    NA    NA     NA    NA    NA    NA
#>   lev_D
#> 1    NA
#> 2    NA
#> 3    NA
#> 
#> $references
#>     glottosubcode reference page var001_ref var001_page var002_ref var002_page
#> 1 tani1257_a_0001        NA   NA         NA          NA         NA          NA
#> 2 tani1257_b_0001        NA   NA         NA          NA         NA          NA
#>   var003_ref var003_page
#> 1         NA          NA
#> 2         NA          NA
#> 
#> $remarks
#>     glottosubcode remark var001_remark var002_remark var003_remark
#> 1 tani1257_a_0001     NA            NA            NA            NA
#> 2 tani1257_b_0001     NA            NA            NA            NA
#> 
#> $contributors
#>     glottosubcode contributor var001_contributor var002_contributor
#> 1 tani1257_a_0001          NA                 NA                 NA
#> 2 tani1257_b_0001          NA                 NA                 NA
#>   var003_contributor
#> 1                 NA
#> 2                 NA
#> 
#> $sample
#>   glottocode group subgroup
#> 1   yucu1253    NA       NA
#> 2   tani1257    NA       NA
#> 
#> $readme
#>                                                        info           value
#> 1                                                maintainer            <NA>
#> 2                                                     email            <NA>
#> 3                                                  citation            <NA>
#> 4                                                       url            <NA>
#> 5 This database was created using the glottospace R package version 0.0.113
#> 
#> $lookup
#>   type_lookup
#> 1        symm
#> 2       asymm
#> 3      factor
#> 4     ordered
#> 5     numeric
#> 6    ordratio
#> 7    logratio
#>                                                                                                                                                   type_legend
#> 1 Symmetric binary: positive match is considered the same as a negative match. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 2          Asymmetric binary: positive match and negative match are not equal. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 3                                    Nominal variables. Levels should be given by a letter or abbreviation. For example: A, B, C. NA stands for missing value
#> 4                                                                                                              Ordinal variables. NA stands for missing value
#> 5                                                                                                Interval scaled variables. Can have any numeric value and NA
#> 6                                               Ratio scaled variables. Setting type to ordratio indicates that values should be treated as ordinal variables
#> 7                                     Ratio scaled variables. Setting type to logratio indicates that values should be log transformed in subsequent analyses
#> 
#> attr(,"class")
#> [1] "glottosubdata" "list"         

# Create glottodata table and add some information to the readme table:
glottocreate(glottocodes = c("yucu1253", "tani1257"), variables = 3,
maintainer = "Your name", email = "yourname@domain.com")
#> $glottodata
#>   glottocode var001 var002 var003
#> 1   yucu1253     NA     NA     NA
#> 2   tani1257     NA     NA     NA
#> 
#> $structure
#>   varname type levels weight group subgroup
#> 1  var001   NA     NA      1    NA       NA
#> 2  var002   NA     NA      1    NA       NA
#> 3  var003   NA     NA      1    NA       NA
#> 
#> $description
#>   varname description reference remarks lev_Y lev_N lev_NA lev_A lev_B lev_C
#> 1  var001          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 2  var002          NA        NA      NA    NA    NA     NA    NA    NA    NA
#> 3  var003          NA        NA      NA    NA    NA     NA    NA    NA    NA
#>   lev_D
#> 1    NA
#> 2    NA
#> 3    NA
#> 
#> $references
#>   glottocode reference page var001_ref var001_page var002_ref var002_page
#> 1   yucu1253        NA   NA         NA          NA         NA          NA
#> 2   tani1257        NA   NA         NA          NA         NA          NA
#>   var003_ref var003_page
#> 1         NA          NA
#> 2         NA          NA
#> 
#> $remarks
#>   glottocode remark var001_remark var002_remark var003_remark
#> 1   yucu1253     NA            NA            NA            NA
#> 2   tani1257     NA            NA            NA            NA
#> 
#> $contributors
#>   glottocode contributor var001_contributor var002_contributor
#> 1   yucu1253          NA                 NA                 NA
#> 2   tani1257          NA                 NA                 NA
#>   var003_contributor
#> 1                 NA
#> 2                 NA
#> 
#> $sample
#>   glottocode group subgroup
#> 1   yucu1253    NA       NA
#> 2   tani1257    NA       NA
#> 
#> $readme
#>                                                        info               value
#> 1                                                maintainer           Your name
#> 2                                                     email yourname@domain.com
#> 3                                                  citation                <NA>
#> 4                                                       url                <NA>
#> 5 This database was created using the glottospace R package     version 0.0.113
#> 
#> $lookup
#>   type_lookup
#> 1        symm
#> 2       asymm
#> 3      factor
#> 4     ordered
#> 5     numeric
#> 6    ordratio
#> 7    logratio
#>                                                                                                                                                   type_legend
#> 1 Symmetric binary: positive match is considered the same as a negative match. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 2          Asymmetric binary: positive match and negative match are not equal. Variable has three possible values: Y, N, NA where NA stands for missing value
#> 3                                    Nominal variables. Levels should be given by a letter or abbreviation. For example: A, B, C. NA stands for missing value
#> 4                                                                                                              Ordinal variables. NA stands for missing value
#> 5                                                                                                Interval scaled variables. Can have any numeric value and NA
#> 6                                               Ratio scaled variables. Setting type to ordratio indicates that values should be treated as ordinal variables
#> 7                                     Ratio scaled variables. Setting type to logratio indicates that values should be log transformed in subsequent analyses
#> 
#> attr(,"class")
#> [1] "glottodata" "list"