Filter glottodata by language, glottocode, etc.

By default, the glottolog data will be used to filter from. But in case the user provides glottodata, this will be used.

glottofilter(
  glottodata = NULL,
  glottocode = NULL,
  location = NULL,
  name = NULL,
  family = NULL,
  family_id = NULL,
  continent = NULL,
  country = NULL,
  sovereignty = NULL,
  macroarea = NULL,
  expression = NULL,
  isocodes = NULL,
  colname = NULL,
  select = NULL,
  drop = NULL
)

Arguments

glottodata

A glottodata table

glottocode

A character vector of glottocodes

location

A character vector with a location (either a continent, country, macroarea, or sovereignty)

name

A character vector of language names

family

A character vector of language families

family_id

A character vector of language family IDs

continent

A character vector of continents

country

A character vector of countries

sovereignty

Sovereignty

macroarea

Glottolog macroarea

expression

A logical expression

isocodes

A character vector of iso639p3codes

colname

A column name

select

Character vector of things to select (only if colname is provided)

drop

Character vector of things to drop (only if colname is provided)

Value

A subset of the original glottodata table (data.frame or sf) containing only filtered languages.

See also

Examples

# \donttest{
points <- glottofilter(location = "Australia")
points <- glottofilter(glottocode = "wari1268")
points <- glottofilter(family = "Indo-European")
points <- glottofilter(continent = "South America")
points <- glottofilter(family = "Indo-European", continent = "South America")
points <- glottofilter(country = c("Colombia", "Venezuela"))
points <- glottofilter(expression = family %in% c("Arawakan", "Tucanoan"))
points <- glottofilter(expression = family_size > 2)
points <- glottofilter(colname = "family", drop = "Indo-European")
# }