Flexible version of glottofilter

This is a more flexible alternative to glottofilter, allowing for negative selections.

glottofilterflex(glottodata, colname, select)

Arguments

glottodata

glottodata table

colname

Character with a single column name

select

Things to be selected in that column

Examples

# \donttest{
glottodata <- glottoget("glottobase")
glottofilterflex(glottodata = glottodata, colname = "family",
select = "Indo-European")
#> Simple feature collection with 582 features and 18 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -155.437 ymin: -29.3255 xmax: 177.772 ymax: 66.32
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    glottocode                       name     macroarea isocode
#> 1    adha1238                     Adhari       Eurasia        
#> 2    adiv1239 Kotia-Adivasi Oriya-Desiya       Eurasia     ort
#> 3    adiw1235            Adiwasi Garasia       Eurasia     gas
#> 4    aerr1238                        Aer       Eurasia     aeq
#> 5    afri1274                  Afrikaans        Africa     afr
#> 6    afro1254       Afro-Seminole Creole North America     afs
#> 7    aima1241                      Aimaq       Eurasia     aiq
#> 8    alvi1241              Alviri-Vidari       Eurasia     avd
#> 9    anci1242              Ancient Greek       Eurasia     grc
#> 10   anda1280       Andaman Creole Hindi       Eurasia     hca
#>            countries family_id
#> 1                 AZ  indo1319
#> 2                 IN  indo1319
#> 3                 IN  indo1319
#> 4                 PK  indo1319
#> 5  BW;MZ;NA;ZA;ZM;ZW  indo1319
#> 6              MX;US  indo1319
#> 7              AF;IR  indo1319
#> 8                 IR  indo1319
#> 9                 GR  indo1319
#> 10                IN  indo1319
#>                                                                                                                                     classification
#> 1                                                                          indo1319/clas1257/indo1320/iran1269/cent2317/cent2318/nort3177/tati1243
#> 2                                                                 indo1319/clas1257/indo1320/indo1321/midd1375/cont1248/indo1323/oriy1254/macr1269
#> 3                                                                 indo1319/clas1257/indo1320/indo1321/midd1375/cont1248/midl1245/bhil1254/gara1268
#> 4                                               indo1319/clas1257/indo1320/indo1321/midd1375/cont1248/midl1245/apab1234/guja1255/guja1256/west2830
#> 5                                                        indo1319/clas1257/germ1287/nort3152/west2793/macr1270/midd1347/mode1257/glob1241/afri1273
#> 6  indo1319/clas1257/germ1287/nort3152/west2793/nort3175/angl1264/angl1265/late1254/merc1242/macr1271/guin1259/cari1284/east2759/gull1242/gull1243
#> 7                                                        indo1319/clas1257/indo1320/iran1269/sout3157/midd1352/mode1259/fars1254/fars1255/east2745
#> 8                                                        indo1319/clas1257/indo1320/iran1269/cent2317/cent2318/nort3177/tati1243/tati1244/sout3177
#> 9                                                                                                     indo1319/clas1257/grae1234/gree1276/east2798
#> 10                                              indo1319/clas1257/indo1320/indo1321/midd1375/cont1248/midl1245/shau1239/indo1322/west2812/hind1270
#>    parent_id        family isolate family_size family_size_rank
#> 1   tati1243 Indo-European   FALSE         583               51
#> 2   macr1269 Indo-European   FALSE         583               51
#> 3   gara1268 Indo-European   FALSE         583               51
#> 4   west2830 Indo-European   FALSE         583               51
#> 5   afri1273 Indo-European   FALSE         583               51
#> 6   gull1243 Indo-European   FALSE         583               51
#> 7   east2745 Indo-European   FALSE         583               51
#> 8   sout3177 Indo-European   FALSE         583               51
#> 9   east2798 Indo-European   FALSE         583               51
#> 10  hind1270 Indo-European   FALSE         583               51
#>                     country              sovereignty              type
#> 1                Azerbaijan               Azerbaijan Sovereign country
#> 2                     India                    India Sovereign country
#> 3                     India                    India Sovereign country
#> 4                  Pakistan                 Pakistan Sovereign country
#> 5                  Zimbabwe                 Zimbabwe Sovereign country
#> 6  United States of America United States of America           Country
#> 7                      Iran                     Iran Sovereign country
#> 8                      Iran                     Iran Sovereign country
#> 9                    Greece                   Greece Sovereign country
#> 10                    India                    India Sovereign country
#>                     geounit     continent adm0_a3                 geometry
#> 1                Azerbaijan          Asia     AZE      POINT (48.06 40.05)
#> 2                     India          Asia     IND  POINT (82.8487 18.1046)
#> 3                     India          Asia     IND  POINT (73.1426 24.0712)
#> 4                  Pakistan          Asia     PAK    POINT (69.029 25.254)
#> 5                  Zimbabwe        Africa     ZWE           POINT (30 -22)
#> 6  United States of America North America     USA POINT (-99.2934 31.7843)
#> 7                      Iran          Asia     IRN   POINT (59.1756 35.283)
#> 8                      Iran          Asia     IRN  POINT (50.3794 35.0075)
#> 9                    Greece        Europe     GRC  POINT (21.9129 39.8155)
#> 10                    India          Asia     IND    POINT (92.738 11.676)
glottofilterflex(glottodata = glottodata, colname = "family",
select = -"Indo-European")
#> Simple feature collection with 7042 features and 18 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -178.785 ymin: -55.2748 xmax: 179.306 ymax: 73.1354
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    glottocode                   name macroarea isocode countries family_id
#> 1    aari1239                   Aari    Africa     aiw        ET  sout2845
#> 2    aasa1238                  Aasax    Africa     aas        TZ  afro1255
#> 3    abad1241                  Abadi Papunesia     kbt        PG  aust1307
#> 4    abag1245                  Abaga Papunesia     abg        PG  nucl1709
#> 5    abai1240            Abai Sungai Papunesia     abf        MY  aust1307
#> 6    abai1241 Abai Tubu-Abai Sembuak Papunesia                ID  aust1307
#> 7    aban1242                Abanyom    Africa     abm        NG  atla1278
#> 8    abar1238                Mungbam    Africa     mij        CM  atla1278
#> 9    abau1245                   Abau Papunesia     aau        PG  sepi1257
#> 10   abaz1241                  Abaza   Eurasia     abq     RU;TR  abkh1242
#>                                                                                        classification
#> 1                                                                          sout2845/ahkk1235/aari1238
#> 2                                                                          afro1255/cush1243/sout3054
#> 3           aust1307/mala1545/cent2237/east2712/ocea1241/west2818/papu1253/peri1258/cent2070/west2850
#> 4                                      nucl1709/kain1273/goro1272/nucl1760/nucl1756/kama1374/unun9976
#> 5                                               aust1307/mala1545/nort3253/sout3154/grea1293/pait1248
#> 6                                               aust1307/mala1545/nort3253/sout3154/grea1294/muru1275
#> 7  atla1278/volt1241/benu1247/bant1294/sout3152/ekoi1237/ekoi1236/ejag1240/bako1254/nort3374/aban1245
#> 8                                               atla1278/volt1241/benu1247/bant1294/sout3152/yemn1234
#> 9                                                                                            sepi1257
#> 10                                                                                  abkh1242/abkh1243
#>    parent_id                   family isolate family_size family_size_rank
#> 1   aari1238             South Omotic   FALSE           5                5
#> 2   sout3054             Afro-Asiatic   FALSE         379               49
#> 3   west2850             Austronesian   FALSE        1271               52
#> 4   unun9976 Nuclear Trans New Guinea   FALSE         317               48
#> 5   pait1248             Austronesian   FALSE        1271               52
#> 6   muru1275             Austronesian   FALSE        1271               52
#> 7   aban1245           Atlantic-Congo   FALSE        1406               53
#> 8   yemn1234           Atlantic-Congo   FALSE        1406               53
#> 9   sepi1257                    Sepik   FALSE          36               27
#> 10  abkh1243             Abkhaz-Adyge   FALSE           5                5
#>                        country                 sovereignty              type
#> 1                     Ethiopia                    Ethiopia Sovereign country
#> 2  United Republic of Tanzania United Republic of Tanzania Sovereign country
#> 3             Papua New Guinea            Papua New Guinea          Geo unit
#> 4             Papua New Guinea            Papua New Guinea          Geo unit
#> 5                     Malaysia                    Malaysia Sovereign country
#> 6                    Indonesia                   Indonesia Sovereign country
#> 7                      Nigeria                     Nigeria Sovereign country
#> 8                     Cameroon                    Cameroon Sovereign country
#> 9             Papua New Guinea            Papua New Guinea          Geo unit
#> 10                      Russia                      Russia Sovereign country
#>             geounit continent adm0_a3                  geometry
#> 1          Ethiopia    Africa     ETH   POINT (36.5721 5.95034)
#> 2          Tanzania    Africa     TZA  POINT (36.8648 -4.00679)
#> 3  Papua New Guinea   Oceania     PNG  POINT (146.992 -9.03389)
#> 4  Papua New Guinea   Oceania     PNG  POINT (145.665 -6.12028)
#> 5          Malaysia      Asia     MYS   POINT (118.306 5.55394)
#> 6         Indonesia      Asia     IDN POINT (116.1625 3.524226)
#> 7           Nigeria    Africa     NGA   POINT (8.63055 6.30619)
#> 8          Cameroon    Africa     CMR    POINT (10.2267 6.5805)
#> 9  Papua New Guinea   Oceania     PNG  POINT (141.324 -3.97222)
#> 10           Russia    Europe     RUS          POINT (42 44.25)
glottofilterflex(glottodata = glottodata, colname = "country",
select = c("Germany", "Netherlands") )
#> Simple feature collection with 31 features and 18 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 3.85428 ymin: 48.254 xmax: 14.4098 ymax: 53.48333
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    glottocode                          name macroarea isocode
#> 1    dutc1256                         Dutch   Eurasia     nld
#> 2    east2290                   Eastphalian   Eurasia        
#> 3    gron1242                      Gronings   Eurasia     gos
#> 4    hind1273 Hindeloopen-Molkwerum Frisian   Eurasia        
#> 5    kols1241                        Kölsch   Eurasia     ksh
#> 6    limb1263                     Limburgan   Eurasia     lim
#> 7    lowe1385                 Lower Sorbian   Eurasia     dsb
#> 8    main1267                  Ostfränkisch   Eurasia     vmf
#> 9    midd1318             Middle Low German   Eurasia     gml
#> 10   midd1321                  Middle Dutch   Eurasia     dum
#>                           countries family_id
#> 1  AW;BE;BQ;BR;CW;DE;GF;GY;NL;SR;SX  indo1319
#> 2                                DE  indo1319
#> 3                             DE;NL  indo1319
#> 4                                NL  indo1319
#> 5                                DE  indo1319
#> 6                          BE;DE;NL  indo1319
#> 7                                DE  indo1319
#> 8                                DE  indo1319
#> 9                                DE  indo1319
#> 10                               NL  indo1319
#>                                                                                        classification
#> 1                    indo1319/clas1257/germ1287/nort3152/west2793/macr1270/midd1347/mode1257/glob1241
#> 2           indo1319/clas1257/germ1287/nort3152/west2793/nort3175/alts1234/midd1345/lowg1239/west2357
#> 3  indo1319/clas1257/germ1287/nort3152/west2793/nort3175/alts1234/midd1345/lowg1239/west2357/ostf1234
#> 4                    indo1319/clas1257/germ1287/nort3152/west2793/nort3175/angl1264/fris1239/mode1264
#> 5           indo1319/clas1257/germ1287/nort3152/west2793/high1289/fran1268/high1287/midd1319/ripu1236
#> 6           indo1319/clas1257/germ1287/nort3152/west2793/high1289/fran1268/high1287/midd1319/ripu1236
#> 7                                               indo1319/clas1257/balt1263/slav1255/west2792/sorb1249
#> 8           indo1319/clas1257/germ1287/nort3152/west2793/high1289/high1286/midd1349/mode1258/uppe1464
#> 9                             indo1319/clas1257/germ1287/nort3152/west2793/nort3175/alts1234/midd1345
#> 10                                     indo1319/clas1257/germ1287/nort3152/west2793/macr1270/midd1347
#>    parent_id        family isolate family_size family_size_rank     country
#> 1   glob1241 Indo-European   FALSE         583               51 Netherlands
#> 2   west2357 Indo-European   FALSE         583               51     Germany
#> 3   ostf1234 Indo-European   FALSE         583               51 Netherlands
#> 4   mode1264 Indo-European   FALSE         583               51 Netherlands
#> 5   ripu1236 Indo-European   FALSE         583               51     Germany
#> 6   ripu1236 Indo-European   FALSE         583               51     Germany
#> 7   sorb1249 Indo-European   FALSE         583               51     Germany
#> 8   uppe1464 Indo-European   FALSE         583               51     Germany
#> 9   midd1345 Indo-European   FALSE         583               51     Germany
#> 10  midd1347 Indo-European   FALSE         583               51 Netherlands
#>    sovereignty              type     geounit continent adm0_a3
#> 1  Netherlands           Country Netherlands    Europe     NLD
#> 2      Germany Sovereign country     Germany    Europe     DEU
#> 3  Netherlands           Country Netherlands    Europe     NLD
#> 4  Netherlands           Country Netherlands    Europe     NLD
#> 5      Germany Sovereign country     Germany    Europe     DEU
#> 6      Germany Sovereign country     Germany    Europe     DEU
#> 7      Germany Sovereign country     Germany    Europe     DEU
#> 8      Germany Sovereign country     Germany    Europe     DEU
#> 9      Germany Sovereign country     Germany    Europe     DEU
#> 10 Netherlands           Country Netherlands    Europe     NLD
#>                    geometry
#> 1              POINT (5 52)
#> 2  POINT (10.00572 52.1179)
#> 3     POINT (6.698 53.2842)
#> 4    POINT (5.4013 52.9427)
#> 5      POINT (6.927 50.947)
#> 6    POINT (5.87705 51.015)
#> 7   POINT (13.9407 51.6621)
#> 8      POINT (8.621 50.005)
#> 9        POINT (9.67 53.14)
#> 10      POINT (5.34 51.656)
glottofilterflex(glottodata = glottodata, colname = "country",
select = -c("Germany", "Netherlands") )
#> Simple feature collection with 7593 features and 18 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -178.785 ymin: -55.2748 xmax: 179.306 ymax: 73.1354
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    glottocode                   name macroarea isocode countries family_id
#> 1    aari1239                   Aari    Africa     aiw        ET  sout2845
#> 2    aasa1238                  Aasax    Africa     aas        TZ  afro1255
#> 3    abad1241                  Abadi Papunesia     kbt        PG  aust1307
#> 4    abag1245                  Abaga Papunesia     abg        PG  nucl1709
#> 5    abai1240            Abai Sungai Papunesia     abf        MY  aust1307
#> 6    abai1241 Abai Tubu-Abai Sembuak Papunesia                ID  aust1307
#> 7    aban1242                Abanyom    Africa     abm        NG  atla1278
#> 8    abar1238                Mungbam    Africa     mij        CM  atla1278
#> 9    abau1245                   Abau Papunesia     aau        PG  sepi1257
#> 10   abaz1241                  Abaza   Eurasia     abq     RU;TR  abkh1242
#>                                                                                        classification
#> 1                                                                          sout2845/ahkk1235/aari1238
#> 2                                                                          afro1255/cush1243/sout3054
#> 3           aust1307/mala1545/cent2237/east2712/ocea1241/west2818/papu1253/peri1258/cent2070/west2850
#> 4                                      nucl1709/kain1273/goro1272/nucl1760/nucl1756/kama1374/unun9976
#> 5                                               aust1307/mala1545/nort3253/sout3154/grea1293/pait1248
#> 6                                               aust1307/mala1545/nort3253/sout3154/grea1294/muru1275
#> 7  atla1278/volt1241/benu1247/bant1294/sout3152/ekoi1237/ekoi1236/ejag1240/bako1254/nort3374/aban1245
#> 8                                               atla1278/volt1241/benu1247/bant1294/sout3152/yemn1234
#> 9                                                                                            sepi1257
#> 10                                                                                  abkh1242/abkh1243
#>    parent_id                   family isolate family_size family_size_rank
#> 1   aari1238             South Omotic   FALSE           5                5
#> 2   sout3054             Afro-Asiatic   FALSE         379               49
#> 3   west2850             Austronesian   FALSE        1271               52
#> 4   unun9976 Nuclear Trans New Guinea   FALSE         317               48
#> 5   pait1248             Austronesian   FALSE        1271               52
#> 6   muru1275             Austronesian   FALSE        1271               52
#> 7   aban1245           Atlantic-Congo   FALSE        1406               53
#> 8   yemn1234           Atlantic-Congo   FALSE        1406               53
#> 9   sepi1257                    Sepik   FALSE          36               27
#> 10  abkh1243             Abkhaz-Adyge   FALSE           5                5
#>                        country                 sovereignty              type
#> 1                     Ethiopia                    Ethiopia Sovereign country
#> 2  United Republic of Tanzania United Republic of Tanzania Sovereign country
#> 3             Papua New Guinea            Papua New Guinea          Geo unit
#> 4             Papua New Guinea            Papua New Guinea          Geo unit
#> 5                     Malaysia                    Malaysia Sovereign country
#> 6                    Indonesia                   Indonesia Sovereign country
#> 7                      Nigeria                     Nigeria Sovereign country
#> 8                     Cameroon                    Cameroon Sovereign country
#> 9             Papua New Guinea            Papua New Guinea          Geo unit
#> 10                      Russia                      Russia Sovereign country
#>             geounit continent adm0_a3                  geometry
#> 1          Ethiopia    Africa     ETH   POINT (36.5721 5.95034)
#> 2          Tanzania    Africa     TZA  POINT (36.8648 -4.00679)
#> 3  Papua New Guinea   Oceania     PNG  POINT (146.992 -9.03389)
#> 4  Papua New Guinea   Oceania     PNG  POINT (145.665 -6.12028)
#> 5          Malaysia      Asia     MYS   POINT (118.306 5.55394)
#> 6         Indonesia      Asia     IDN POINT (116.1625 3.524226)
#> 7           Nigeria    Africa     NGA   POINT (8.63055 6.30619)
#> 8          Cameroon    Africa     CMR    POINT (10.2267 6.5805)
#> 9  Papua New Guinea   Oceania     PNG  POINT (141.324 -3.97222)
#> 10           Russia    Europe     RUS          POINT (42 44.25)
glottofilterflex(glottodata = glottodata, colname = "continent",
select = -c("South America", "Europe", "Asia", "Oceania", "Africa") )
#> Simple feature collection with 941 features and 18 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -176.207 ymin: -55.2748 xmax: 179.1981 ymax: 69.582
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    glottocode                 name     macroarea isocode countries family_id
#> 1    acat1239     Acatepec Me'phaa North America     tpx        MX  otom1299
#> 2    achi1256                 Achi North America     acr        GT  maya1287
#> 3    achu1247             Achumawi North America     acv        US  pala1350
#> 4    adai1235                 Adai North America     xad        US  adai1235
#> 5    afro1254 Afro-Seminole Creole North America     afs     MX;US  indo1319
#> 6    agua1252           Aguacateco North America     agu        GT  maya1287
#> 7    ahte1237               Ahtena North America     aht        US  atha1245
#> 8    aikl1237               Aiklep     Papunesia     mwg        PG  aust1307
#> 9    akei1237                 Akei     Papunesia     tsr        VU  aust1307
#> 10   alab1237              Alabama North America     akz        US  musk1252
#>                                                                                                                                     classification
#> 1                                                                                   otom1299/west2783/tlap1252/subt1249/meph1234/nort3350/west2948
#> 2                                                                                            maya1287/core1254/quic1274/grea1276/core1251/quic1275
#> 3                                                                                                                                         pala1350
#> 4                                                                                                                                             <NA>
#> 5  indo1319/clas1257/germ1287/nort3152/west2793/nort3175/angl1264/angl1265/late1254/merc1242/macr1271/guin1259/cari1284/east2759/gull1242/gull1243
#> 6                                                                                                     maya1287/core1254/quic1274/grea1277/ixil1250
#> 7                                                                                                              atha1245/atha1246/atha1247/sout3337
#> 8                             aust1307/mala1545/cent2237/east2712/ocea1241/west2818/nort3206/nger1241/viti1243/sout2874/araw1269/araw1283/west2530
#> 9                                                        aust1307/mala1545/cent2237/east2712/ocea1241/nort3195/nort3205/espi1234/west2956/west2845
#> 10                                                                                                                               musk1252/alab1247
#>    parent_id                  family isolate family_size family_size_rank
#> 1   west2948             Otomanguean   FALSE         181               46
#> 2   quic1275                   Mayan   FALSE          33               25
#> 3   pala1350             Palaihnihan   FALSE           2                2
#> 4       <NA>                    Adai    TRUE           1                1
#> 5   gull1243           Indo-European   FALSE         583               51
#> 6   ixil1250                   Mayan   FALSE          33               25
#> 7   sout3337 Athabaskan-Eyak-Tlingit   FALSE          46               34
#> 8   west2530            Austronesian   FALSE        1271               52
#> 9   west2845            Austronesian   FALSE        1271               52
#> 10  alab1247               Muskogean   FALSE           7                7
#>                     country              sovereignty              type
#> 1                    Mexico                   Mexico Sovereign country
#> 2                 Guatemala                Guatemala Sovereign country
#> 3  United States of America United States of America           Country
#> 4  United States of America United States of America           Country
#> 5  United States of America United States of America           Country
#> 6                 Guatemala                Guatemala Sovereign country
#> 7  United States of America United States of America           Country
#> 8                      <NA>                     <NA>              <NA>
#> 9                      <NA>                     <NA>              <NA>
#> 10 United States of America United States of America           Country
#>                     geounit     continent adm0_a3                   geometry
#> 1                    Mexico North America     MEX   POINT (-99.0602 17.1034)
#> 2                 Guatemala North America     GTM    POINT (-90.3513 15.073)
#> 3  United States of America North America     USA   POINT (-121.191 41.0979)
#> 4  United States of America North America     USA   POINT (-92.0059 31.6314)
#> 5  United States of America North America     USA   POINT (-99.2934 31.7843)
#> 6                 Guatemala North America     GTM   POINT (-91.3153 15.3307)
#> 7  United States of America North America     USA   POINT (-144.889 61.7068)
#> 8                      <NA>          <NA>    <NA>   POINT (149.571 -6.23575)
#> 9                      <NA>          <NA>    <NA>   POINT (166.767 -15.6377)
#> 10 United States of America North America     USA POINT (-94.62848 30.76738)
# }