glottomatch

Match a vector of language names to glottocodes and names

glottomatch(namevec, glottodata = NULL, tolerance = NULL)

Arguments

namevec

Vector of language names

glottodata

Optional, where to search for matches. If kept empty, the entire glottolog database will be searched, you could also search within a specific area

tolerance

Optional, search tolerance.

Value

a data.frame with exact or closest matches, and their glottocodes.

Examples

glottodata <- glottofilter(continent = "South America")
# Finds a single match
glottomatch(name = "yucuni", glottodata = glottodata)
#>     name matchname glottocode
#> 1 yucuni    Yucuna   yucu1253
# Finds multiple matches
glottomatch(name = "quechui", glottodata = glottodata)
#>       name                             matchname glottocode
#> 1  quechui                    Ambo-Pasco Quechua   ambo1252
#> 2  quechui             Arequipa-La Unión Quechua   areq1237
#> 3  quechui                      Ayacucho Quechua   ayac1239
#> 4  quechui                     Cajamarca Quechua   caja1238
#> 5  quechui          Cajatambo North Lima Quechua   caja1239
#> 6  quechui                   Chachapoyas Quechua   chac1250
#> 7  quechui                       Chincha Quechua   chin1483
#> 8  quechui               Chiquián Ancash Quechua   chiq1249
#> 9  quechui                     Classical Quechua   clas1251
#> 10 quechui                Corongo Ancash Quechua   coro1247
#> 11 quechui                         Cusco Quechua   cusc1236
#> 12 quechui              Eastern Apurímac Quechua   east2551
#> 13 quechui              Huallaga Huánuco Quechua   hual1241
#> 14 quechui Huamalíes-Dos de Mayo Huánuco Quechua   huam1248
#> 15 quechui                Huaylas Ancash Quechua   huay1240
#> 16 quechui                 Huaylla Wanca Quechua   huay1241
#> 17 quechui                   Jauja Wanca Quechua   jauj1238
#> 18 quechui                    Lambayeque Quechua   lamb1276
#> 19 quechui   Margos-Yarowilca-Lauricocha Quechua   marg1254
#> 20 quechui                  Napo Lowland Quechua   napo1242
#> 21 quechui                North Bolivian Quechua   nort2976
#> 22 quechui     Northern Conchucos Ancash Quechua   nort2979
#> 23 quechui                   North Junín Quechua   nort2980
#> 24 quechui                      Pacaraos Quechua   paca1245
#> 25 quechui                 Panao Huánuco Quechua   pana1306
#> 26 quechui                          Puno Quechua   puno1238
#> 27 quechui                    San Martín Quechua   sanm1289
#> 28 quechui                 Sihuas Ancash Quechua   sihu1235
#> 29 quechui              Southern Pastaza Quechua   sout2990
#> 30 quechui                South Bolivian Quechua   sout2991
#> 31 quechui     Southern Conchucos Ancash Quechua   sout2992
#> 32 quechui                Chaupihuaranga Quechua   yana1272
#> 33 quechui                        Yauyos Quechua   yauy1235