Save glottodata, maps and plots

If no filename is provided, the name of the glottodata object will be used.

glottosave(glottodata, filename = NULL)

Arguments

glottodata

User-provided glottodata

filename

Filename either with or without file extension

Value

No object is returned, it will be save locally at the specified location

Details

If no file extension is provided, a sensible default file extension is chosen. Dynamic maps (tmap) are saved in .html format, static maps (tmap) are saved as .png. Spatial data (sf) are saved as geopackage (.GPKG) by default, but .shp is also possible.

See also

Examples

# \donttest{
glottodata <- glottoget("demodata", meta = FALSE)
# Saves as .xlsx
glottosave(glottodata, filename = file.path(tempdir(), "glottodata") )
#> Glottodata (glottodata) saved as /tmp/Rtmp8v6osF/glottodata.xlsx

glottospacedata <- glottospace(glottodata)
# Saves as .GPKG
glottosave(glottospacedata, filename = file.path(tempdir(), "glottodata") )
#> Writing layer `glottodata' to data source 
#>   `/tmp/Rtmp8v6osF/glottodata.gpkg' using driver `GPKG'
#> Writing 6 features with 10 fields and geometry type Point.
#> Spatial object (sf) saved as /tmp/Rtmp8v6osF/glottodata.gpkg

glottomap <- glottomap(glottodata)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
# Saves as .png
glottosave(glottomap, filename = file.path(tempdir(), "glottomap") )
#> Map (tmap object) saved as /tmp/Rtmp8v6osF/glottomap.png

# Saves as .html
glottomap <- glottomap(glottodata, type = "dynamic",
             filename = file.path(tempdir(), "glottomap") )
#> 
#> ── tmap v3 code detected ───────────────────────────────────────────────────────
#> Map (tmap object) saved as /tmp/Rtmp8v6osF/glottomap.html
# }