Permanova across all groups (overall or pairwise)

This function takes a glottodata or glottosubdata object and performs a Permutational Multivariate Analysis of Variance (PERMANOVA). It can be used to test whether two or more groups are significantly different from each other (by specifying the 'comparison' argument with either 'overall' or 'pairwise'). The function uses the 'group' column in the sample table to do the comparisons. Before running the analysis, a distance matrix is constructed from the glotto(sub)data object using glottodist(). The function calls vegan::adonis2(), type ?adonis2 for more details.

glottostat_permanova(
  glottodata,
  comparison = NULL,
  sample = NULL,
  permutations = NULL,
  metric = "gower"
)

Arguments

glottodata

glottodata or glottosubdata

comparison

Either "overall" or "pairwise"

sample

sample table (optional). By default, searches for sample table in glottodata/glottosubdata.

permutations

Number of permutations (default is 999)

metric

Either "gower" or "anderberg"

Examples

glottodata <- glottoget("demodata", meta = TRUE)
glottostat_permanova(glottodata, comparison = "pairwise")
#> glottocode used as id
#> glottocode used as id
#> Out of the initial 6 data points, 0 have been removed because of missing data. 
#>  6 remaining data points.
#> Running pairwise permanova
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#>   group1 group2 p-value significance p-value (adj) sign (adj)
#> 1      A      B     0.3         n.s.           0.3       n.s.

# Use subgroup (or another column in the structure table) as group
glottodata[["sample"]][,"group"] <- NULL # delete old 'group' column
glottodata[["sample"]][,"group"] <- glottodata[["sample"]][,"subgroup"]
glottostat_permanova(glottodata, comparison = "pairwise")
#> glottocode used as id
#> glottocode used as id
#> Out of the initial 6 data points, 0 have been removed because of missing data. 
#>  6 remaining data points.
#> Running pairwise permanova
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#> 'nperm' >= set of all permutations: complete enumeration.
#> Set of permutations < 'minperm'. Generating entire set.
#>   group1 group2 p-value significance p-value (adj) sign (adj)
#> 1      A      B       1         n.s.             1       n.s.
#> 2      A      C       1         n.s.             1       n.s.
#> 3      A      D       1         n.s.             1       n.s.
#> 4      B      C       1         n.s.             1       n.s.
#> 5      B      D       1         n.s.             1       n.s.
#> 6      C      D      NA         <NA>            NA       <NA>

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottostat_permanova(glottodata = glottosubdata, comparison = "pairwise")
#> glottosubcode used as id
#> glottosubcode used as id
#> Out of the initial 20 data points, 0 have been removed because of missing data. 
#>  20 remaining data points.
#> Running pairwise permanova
#>   group1 group2 p-value significance p-value (adj) sign (adj)
#> 1      A      B   0.714         n.s.         0.714       n.s.